Improve prediction cache.
Since we are always asking for rules with a specific next element
when we deal with the predictions we can generate a better data
structure and only have to calculate that cost once. It also saves
a lot of time of looping over useless entries (that do not match
on the next element).