Try to correctly handle var.
Since var can be either an Identifier or a var literal and the lexer
can not know we now return a set of possible tokens and advance the
matching rules (similar to how error correction already worked).
This meant that TypeIdentifier had to be made a token as well, instead
of just tagging along on Identifier and later on validating itself.