More work to get signatures correct.
CompiledTypesHolder now returns FullNameHandlers based on the types
which means that we can get full type information from the actual classes.
This means that ClassSetter need to set the types from CompiledTypesHolder.
Move TypeParameter signature handling to GenericTypeHelper so that we can
use it for signatures in both classes and methods.